feat: ユーザープロフィールに 実績・リアクション タブを追加
- 実績グリッドを MkAchievementsGrid として共通化し、既存の実績カラムと新しいプロフィール実績タブの両方から再利用 (DeckAchievementsColumn は 733 → 94 行に縮小)
- リアクションタブは publicReactions プライバシーフラグに応じて動的に表示し、users/reactions をページング取得
- useEditorTabs を readonly T[] と Ref<readonly T[]> の両方を受け付けられるよう拡張し、プライバシー設定でタブ集合が変わるケースに対応
