+++ src/composables/useEditorTabs.ts
@@ -1,4 +1,4 @@
-import { computed, type Ref, ref } from 'vue'
+import { computed, isRef, type Ref, ref } from 'vue'
 import { useSwipeTab } from '@/composables/useSwipeTab'
 import { useTabSlide } from '@/composables/useTabSlide'
 
@@ -7,28 +7,36 @@ import { useTabSlide } from '@/composables/useTabSlide'
  *
  * Wraps `useTabSlide` + `useSwipeTab` so each editor doesn't repeat
  * the same boilerplate.
+ *
+ * Pass a plain array for a fixed tab set, or a `Ref<readonly T[]>` when
+ * the visible tabs depend on runtime state (e.g. privacy flags).
  */
 export function useEditorTabs<T extends string>(
-  tabs: readonly T[],
+  tabs: readonly T[] | Ref<readonly T[]>,
   defaultTab: T,
 ) {
+  const tabsRef = computed<readonly T[]>(() =>
+    isRef(tabs) ? tabs.value : tabs,
+  )
   const tab = ref(defaultTab) as Ref<T>
   const containerRef = ref<HTMLElement | null>(null)
-  const tabIndex = computed(() => tabs.indexOf(tab.value))
+  const tabIndex = computed(() => tabsRef.value.indexOf(tab.value))
 
   useTabSlide(tabIndex, containerRef)
 
   useSwipeTab(
     containerRef,
     () => {
-      const next = tabs[tabs.indexOf(tab.value) + 1]
+      const list = tabsRef.value
+      const next = list[list.indexOf(tab.value) + 1]
       if (next) {
         tab.value = next
         return true
       }
     },
     () => {
-      const prev = tabs[tabs.indexOf(tab.value) - 1]
+      const list = tabsRef.value
+      const prev = list[list.indexOf(tab.value) - 1]
       if (prev) {
         tab.value = prev
         return true
